backend/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.33.0
playwright==1.48.0
psycopg2-binary==2.9.10
colorthief==0.2.1
Pillow==10.4.0
httpx==0.28.1
requests==2.22.0
python-multipart==0.0.20
